Implement deleting a tour
